+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Until May 2022 (inclusive) no changelog was kept. We might try to reconstruct it
 * APFEL: Upgrade to 3.1.1 and switch to CMake build
 * ROOT, Geant4: Build with Ninja
 * FairLogger: Update to v2.3.1
+* yaml-cpp: Update to 0.8.0, remove obsolete boost dependency
 * FairShip: Build with Ninja
 * Refactor modulefiles to use alibuild-generate-module
 * FairRoot: Remove unused FairMQ dependency